# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TJ 노래방 곡 검색 API",
        "title": "Musicon API",
        "contact": {},
        "version": "1.0"
    },
    "host": "158.179.160.120:7847",
    "basePath": "/",
    "paths": {
        "/api/songs/search": {
            "get": {
                "description": "제목, 가수, 초성으로 곡을 검색합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "곡 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "검색어 (제목/가수/초성)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "결과 수 (기본 20, 최대 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "오프셋 (기본 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/songs/{number}": {
            "get": {
                "description": "TJ 노래방 번호로 곡을 조회합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "TJ 번호로 곡 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TJ 노래방 번호",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "서버 상태 확인",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "헬스체크",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}